379.2. Creating Context Objects(Section 27: The Context System with React)
So now that we have created our context object we are going to get some information into the object

by setting up a default value.

We can create a default value by passing it directly into the create context call.

So if you create context right here I'm going to pass a string of English like so.

So that's going to indicate to my two components are my two nested children of the field and a button

that by default the selected language that we should be showing on the screen is English.

So the default value for the Context object is this string English now that we have put a default value

into the Context object we're going to connect it to our button component.

First we're going to do the same process for the field but we'll start with the button we're going to

connect that context object to our nested child of button by making use of this this context property.

So we're going to making use of this context to get information out of the Context object rather than

creating this consumer component thing.

Over here we'll talk about why we will use a consumer component later on.

OK.

So inside of my components directory I will find my button dodgiest file and then at the top I'm going

to import the language context object that we just created.

So I'm going to import language context from up one directory contex.

Make sure you get the on there if you spelt the directory name over here with an S as well.

And then I'm going to import the language context file like so.

Now to hook up the language context object to my button component.

Let's look at a quick diagram.

All right so here's what we're going to do.

We're going to first set up something called a context type that is going to essentially link our button

component to the Context object.

And then to actually reference the information that is inside of the Context object.

We're going to make use of the this context property.

So again the first thing we do is set up this context type property back inside of my component file.

I'm going to add a little bit of space inside of my component class and then I'll add in static context

type equals language context like so so this is how we Hookup a context object to a class component.

We say static context type and then we just print out the Context object itself that we had created

inside that other file.

Now remember the context type property here much like many other properties we have created inside of

Riak components is a very special property name.

So if you call this anything else this context stuff is not going to work correctly if you called it

something like my context right.

Nope not going to work.

Ask to be called exactly contex type like so.