379.3. Creating Context Objects(Section 27: The Context System with React)
Now you may have not ever seen the static keyword syntax before.

Essentially what static context typewrite here does is it adds a property to our class itself.

Remember any time that we create a instance of a component by writing out something like button inside

of GSX that is creating an instance of the Button class and each time we add methods to a class or any

time that we update say or state property or anything like that we are changing properties that are

assigned to an instance of our class when we make use of the static.

You were right here.

We are adding a property directly to the class itself.

So the line of code that you see right there is identical to writing out button Dom context type equals

language context.

You can set up a context type property by making use of either methods that you see right here so we

can write out either static contex type inside of the class declaration or after creating the class.

We can then write out the class name Dom context type and then hook up the context like so you can do

it either way.

Totally up to you.