379.4. Creating Context Objects(Section 27: The Context System with React)
Personally I make like making use of the static context type because it puts that declaration at the

very top of the file.

Very easy to read.

Very easy to find.

So now that we have added a reference of context type to our context object to our component class our

component class can now reference this context and get access to the data inside of that context object

which in this case is this string English.

So to make sure that's the case I'm going to find my render method.

And inside there are going to cons. log out this duck context like so.

All right so now it's time to test this out.

I'm going to save the button dachas file.

I'm going to save the language context.

Yes file and then I'll flip back over to my browser.

I'm going to make sure I refresh the page and then when I do so I should see a console log over here

of this string English.

All right so that's pretty much it.

That is a very basic example.

We have now created a context object.

We got a default value into it by passing the default value into the context object when we created

it.

And then we got that default value out of that context object by setting up the context type property

to link our component to the context object.

And we got reference to the actual data inside the context object with this context.

Now one thing I want to kind of clear here is that I'm giving you like a super detailed academic definition

of what's going on.

So if it feels like this is kind of confusing I want to tell you it's really not you know at the end

the day.

Let's just like really simplify this create the context object define context type boom.

You get this context.

That's it that's all that's really going on.

All right.

Now let's going to show you very quickly is that we can very easily change the string that we make use

of when we create our context to change the default value inside that context object.

So for example rather than passing in a string of English I could say simply Dutch.

So if I change that and then save the file we can flip back over NLC a console log if Dutch We do not

have to pass in strings to a context object.

For values that get shared we can use any type of valid javascript value so we can do an object with

a color of red.

So I can save that and I'll see color of red.

We can do an array that has one 2 three.

So I can save that and see one 2 three over here.

So I think you get the idea pretty quickly.

We can share any information that we want inside of our context object.

But for our application it makes sense to have a default value of English.

OK so now that we've got a very simple example put together.

Let's take a quick pause.

When we come back in the next section we're going to make sure that our button component updates itself

based upon that context property.

And then after that we'll make sure that we repeat the same process for our field component as well.

So quick pulse and I'll see you in just a minute.