382.1. Gotchas Around Providers(Section 27: The Context System with React)
In last video we got our application working by implementing this provider component.

So in this video I want to focus on this provider and talk about exactly what it does.

First off inside my language context not just file I went to cancel log out the object that gets returned

from calling create context.

So in front of create context I'm going to replace export default with Konst context or then do a console

log of that value and then I will actually export it.

So export default context like the old and save this and flip back over to my application.

So here is the console log right here.

You'll notice that the Context object that we created has a provider property on it that provider property

is a re-act component.

So back inside of our apts J.S. File when we referenced language context not provider.

This was a real component.

It was created for us automatically by the context system.

You'll also notice that there is a consumer property inside there as well.

You might recall that we very briefly discuss the consumer as being another way that we can consume

information out of the Context object.