382.3. Gotchas Around Providers(Section 27: The Context System with React)
All right so now that we see where that gets created I want to undo that little change we just made.

So going to change us back to being export default

create context like so.

And I want to continue telling you a little bit more about that provider element right.

Here we go.

So our parent component renders that context provider component.

And in this case you'll notice I abbreviated like language context simply as context.

Now the trick here is understanding that whatever value we assign to the value prop of that component

will be used to update our context object.

So for example if we assign the string English to the value prop our context object will be updated

to contain the value.

English as well.

There's no limit to the type of value that we can assign to the value prop.

So this could just as easily be an array of 1 2 3 and that would update the Context object with an array

of 1 2 3 as well.

The next thing I want you to understand about that provider is that we are not limited to always assigning

it a value from this outstate dealt language.

We do not have to make use of the state system when setting up this context stuff.

So we didn't have to use the value from state language here but it just happened that that is how we

put our application together as we just saw in that diagram.

This could have just as easily been English.

But of course that would have hardcoded the value of English to exist inside of our context object which

is probably not what we want.

With that in mind usually you will have a value prop coming from a props object or a state object or

something similar because you usually want to change the value inside of your length or excuse me inside

of your context object over time.