383.2. Accessing Data with Consumers(Section 27: The Context System with React)
You could always define a helper method on your class so I could do something like.

I don't know how about render submit.

And then we could call this with our value and then we can return the result of value equals English

Samit or vore leggin like so and then we could call that from this inner function.

So I could call this render submit and pass in the value like so all right let's try saving this and

see what happens.

So I'll save the file I'm going to flip back over to my application.

I still see the default of name and submit and then if I click on the other flag I'll see the translation

as well.

So again this is another way of getting information out of one of these context objects.

Very similar to this context.

Remember when we use the context approach we also have to set up that context type but we do not have

to do that if we are making use of the consumer approach.

Now the most obvious question that you probably have right now is why would we do this.

Why would we take this approach as opposed to using that context type approach.

Well let's take a quick pause right here.

When we come back the next section I'll tell you why we would want to use a consumer in place of that

context type approach.

So quick break and I'll see you in just a minute.