384.1. Pulling From Multiple Contexts(Section 27: The Context System with React)
In the last video we saw the usage of the consumer component to get information out of a Context object.

So remember that is the second way that we can use to get information out of context.

Now the most obvious question about the consumer is why would we use this as opposed to just using this

context which looked a little bit more straightforward.

Well the answer is that we will make use of consumer any time that we want to get information out of

multiple different context objects inside of a single component using this context in conjunction with

that context type property is only used any time we are accessing a single context object inside a component.

So let's think up an example where we might have multiple different contexts objects so we want to get

information from inside of our buttons.

JS file here's my idea.

I think that we should create a second context object called Color context and this will have a value

of a string read if the value is read.

Then we're going to want to have our button style itself as a red button as opposed to the default blue

that it is right now.

So that means that we're going to need to set up a second Context object.

We will set up a second provider and then we'll try to access the second context object inside of the

bun.

And that's going to be in addition to the current language context that we are already accessing inside

there.

So this will be an example where we are trying to access multiple different contexts objects because

we are accessing multiple different contexts objects.

We will have to use the consumer for both of them.

We cannot use this context in this case so let's try this out.

I'm going to flip back over to my code editor and inside my contex directory.

I will make a new file a file called Color context.

Yes.

So inside of year we are going to create a new context object and export it.

So I will import re-act from react and then I will export default re-act not create context like so

now you do not always have to assign a default value if you want to.

You can here by passing it as an argument to create context.

So in this case we might want a default value of like red or something like that.

But just to be clear we are going to set up that provider that's going to stuff a value into the Context

object anyways.

So we don't have to really provide a default value here because we're going to get a value into it by

using that provider.

OK.

So there's our color context.