384.2. Pulling From Multiple Contexts(Section 27: The Context System with React)
Now we need to wire this thing up inside of our outdoor J.S. file inside of aptest Yes at the very top.

I will import color context from dot dot slash contex color context.

I'm then going to find my render method where we show user create right here.

So I want to make some data available to the button component that gets rendered inside of user create.

So I'm going to wrap user create with a new color context provider.

It doesn't matter if we put it outside like right here or inside of the language context no difference

either way.

We just need to make sure that we somehow wrap user create with or color context provider.

In this case I will place it outside just for fun no particular reason.

So out here I will place color contacts dot provider and I'll set the value to be read.

So in this case we're going to hard code the value as read.

There's going to be no decision making here on what value we put into the context object.

I'll then close off the color context provider Whoops.

Where's my daughter.

Here it is right there down there and then I'm going to fix up the indentation like so.

So we now have two different context objects that can provide information to the user create in every

single component inside of there.

So I'm not going to save this and then I will open up the button not just file.

Now inside here we want to import the color contex we just created and we're going to use its consumer

component to access the value inside of it and change a class name on the button component right here.

So we if we want to show the button as red we will replace primary with the string red instead.

So if I say that really quickly just as a quick example you'll see the button appears as read over here.

So essentially we want to access that context value and decide either to put in primary or read right

here.

So at the top I'm going to import color context from dot dot slash contex color context and then because

I want to access the value that context object and change some property of the button.

I'm going to wrap the button in this case with color context.

Consumer.