384.3. Pulling From Multiple Contexts(Section 27: The Context System with React)
Now here's the part that gets a little bit nasty.

Remember the only child that we passed to a consumer component is going to be a function.

So we now need to pass a function as the only child to this consumer and that function is going to receive

the current value inside of it.

In this case read and then from that function we're going to have to return this entire block of GSX

right here.

So it's going to be a little bit nasty syntax but we can figure it out.

So I'm going to add in a child here that the set of curly braces.

This is going to be a function I'm going to give myself a little bit of space.

And I'm going to cut and paste the button and everything inside of it into that function like so.

So now this function will be called with whatever value is inside of our color context object I could

refer to that argument as value.

But you'll notice we already have a variable inside if you're of value as well.

So just to be a little bit more clear I would probably call this something like color because that's

what we are really receiving here receiving the color that we're supposed to use for our button.

So now I'm going to use a little bit of yes 20:15 syntax to get that color into the class name string

right there.

So I'm going to wrap the thing with curly braces.

I'll then put in a set of tactics around the string and then I will replace primary with Dollar Sign

curly braces color like so.

So we're now taking whatever string is inside of our color context and using it as a class name for

our button.

Let's now save this and see what happens.

So I can refresh the application and I see that I get a button of Samit or a red button right there

and I can still click on the different flags and see the translation as well.

If I go back over to the outtalk J.S. file I can change the value of red right there over to primary

and I will see the button go back over to being blue like it was before.

So you can imagine that if we had like some type of color selector here in addition to the language

selector we could very easily hook up that color selector to our component state and of the Epcot J

s component and have that state property decide either to show the text primary or red or whatever we

need to show.

OK.

So that's pretty much it.

That is how we consume multiple different contexts inside of a single component.

Now as you can see the syntax here is a little bit nasty because we are essentially stacking up these

different consumers and each one needs to have as the only argument a function.

And so like I said yeah gets kind of nasty.